+++ great_expectations/profiler/domain_builder/simple_semantic_type_domain_builder.py
@@ -1,4 +1,4 @@
-from typing import List, Optional, Union
+from typing import Any, Dict, List, Optional, Union, cast
 
 import great_expectations.exceptions as ge_exceptions
 from great_expectations.core.domain_types import MetricDomainTypes, SemanticDomainTypes
@@ -9,10 +9,7 @@ from great_expectations.profiler.domain_builder.domain import Domain
 from great_expectations.profiler.domain_builder.inferred_semantic_domain_type import (
     InferredSemanticDomainType,
 )
-from great_expectations.profiler.util import (
-    infer_semantic_domain_type_from_table_column_type,
-    parse_semantic_domain_type_argument,
-)
+from great_expectations.profiler.util import parse_semantic_domain_type_argument
 from great_expectations.validator.validator import MetricConfiguration, Validator
 
 
@@ -74,7 +71,7 @@ class SimpleSemanticTypeColumnDomainBuilder(ColumnDomainBuilder):
         semantic_domain_type: SemanticDomainTypes
         for column_name in table_column_names:
             inferred_semantic_domain_type = (
-                infer_semantic_domain_type_from_table_column_type(
+                self.infer_semantic_domain_type_from_table_column_type(
                     validator=validator, column_name=column_name
                 )
             )
@@ -92,3 +89,99 @@ class SimpleSemanticTypeColumnDomainBuilder(ColumnDomainBuilder):
                 )
 
         return domains
+
+    # This method (default implementation) can be overwritten (with different implementation mechanisms) by subclasses.
+    # noinspection PyMethodMayBeStatic
+    def infer_semantic_domain_type_from_table_column_type(
+        self, validator: Validator, column_name: str
+    ) -> InferredSemanticDomainType:
+        # Note: As of Python 3.8, specifying argument type in Lambda functions is not supported by Lambda syntax.
+        column_types_dict_list: List[Dict[str, Any]] = list(
+            filter(
+                lambda column_type_dict: column_name in column_type_dict,
+                validator.get_metric(
+                    metric=MetricConfiguration(
+                        metric_name="table.column_types",
+                        metric_domain_kwargs={},
+                        metric_value_kwargs={
+                            "include_nested": True,
+                        },
+                        metric_dependencies=None,
+                    )
+                ),
+            )
+        )
+        if len(column_types_dict_list) != 1:
+            raise ge_exceptions.ProfilerExecutionError(
+                message=f"""Error: {len(column_types_dict_list)} columns were found while obtaining semantic type \
+    information.  Please ensure that the specified column name refers to exactly one column.
+    """
+            )
+
+        column_type: str = cast(str, column_types_dict_list[0][column_name]).upper()
+
+        semantic_column_type: SemanticDomainTypes
+        if column_type in (
+            set([type_name.upper() for type_name in ProfilerTypeMapping.INT_TYPE_NAMES])
+            | set(
+                [
+                    type_name.upper()
+                    for type_name in ProfilerTypeMapping.FLOAT_TYPE_NAMES
+                ]
+            )
+        ):
+            semantic_column_type = SemanticDomainTypes.NUMERIC
+        elif column_type in set(
+            [type_name.upper() for type_name in ProfilerTypeMapping.STRING_TYPE_NAMES]
+        ):
+            semantic_column_type = SemanticDomainTypes.TEXT
+        elif column_type in set(
+            [type_name.upper() for type_name in ProfilerTypeMapping.BOOLEAN_TYPE_NAMES]
+        ):
+            semantic_column_type = SemanticDomainTypes.LOGIC
+        elif column_type in set(
+            [type_name.upper() for type_name in ProfilerTypeMapping.DATETIME_TYPE_NAMES]
+        ):
+            semantic_column_type = SemanticDomainTypes.DATETIME
+        elif column_type in set(
+            [type_name.upper() for type_name in ProfilerTypeMapping.BINARY_TYPE_NAMES]
+        ):
+            semantic_column_type = SemanticDomainTypes.BINARY
+        elif column_type in set(
+            [type_name.upper() for type_name in ProfilerTypeMapping.CURRENCY_TYPE_NAMES]
+        ):
+            semantic_column_type = SemanticDomainTypes.CURRENCY
+        elif column_type in set(
+            [type_name.upper() for type_name in ProfilerTypeMapping.IDENTITY_TYPE_NAMES]
+        ):
+            semantic_column_type = SemanticDomainTypes.IDENTITY
+        elif column_type in (
+            set(
+                [
+                    type_name.upper()
+                    for type_name in ProfilerTypeMapping.MISCELLANEOUS_TYPE_NAMES
+                ]
+            )
+            | set(
+                [
+                    type_name.upper()
+                    for type_name in ProfilerTypeMapping.RECORD_TYPE_NAMES
+                ]
+            )
+        ):
+            semantic_column_type = SemanticDomainTypes.MISCELLANEOUS
+        else:
+            semantic_column_type = SemanticDomainTypes.UNKNOWN
+
+        inferred_semantic_column_type: InferredSemanticDomainType = (
+            InferredSemanticDomainType(
+                semantic_domain_type=semantic_column_type,
+                details={
+                    "algorithm_type": "deterministic",
+                    "mechanism": "lookup_table",
+                    "source": "great_expectations.profile.base.ProfilerTypeMapping",
+                },
+            )
+        )
+
+        return inferred_semantic_column_type
